Crate wasmtime_internal_jit_debug

⚠️ Warning ⚠️: this crate is an internal-only crate for the Wasmtime project and is not intended for general use. APIs are not strictly reviewed for safety and usage outside of Wasmtime may have bugs. If you’re interested in using this feel free to file an issue on the Wasmtime repository to start a discussion about doing so, but otherwise be aware that your usage of this crate is not supported.

Modules

gdb_jit_int

The GDB’s JIT compilation interface. The low level module that exposes the __jit_debug_register_code() and __jit_debug_descriptor to register or unregister generated object images with debuggers.

perf_jitdump

Support for jitdump files which can be used by perf for profiling jitted code. Spec definitions for the output format is as described here: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/tools/perf/Documentation/jitdump-specification.txt